+++ deps.ts
@@ -11,5 +11,5 @@
 
 // Ana utils
 export * from 'https://deno.land/x/ana_utils@v0.0.4/mod.ts';
-export * from 'https://deno.land/x/ana_components@v0.0.38/mod.ts';
+export * from 'https://deno.land/x/ana_components@v0.0.39/mod.ts';
 // export * from '../ana-components/mod.ts';